+++ apps/api/src/modules/farms/farms.module.ts
@@ -1,8 +1,10 @@
 import { Module } from '@nestjs/common';
 import { FarmsService } from './farms.service';
 import { FarmsResolver } from './farms.resolver';
+import { PrismaModule } from '../../prisma/prisma.module';
 
 @Module({
+  imports: [PrismaModule],
   providers: [FarmsService, FarmsResolver],
   exports: [FarmsService],
 })